requirements.txt
Pillow
requests
torch
torchvision
tqdm
nltk
